Confilict

+++ app/Http/Controllers/Admin/DataGuruController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Http\Requests\Admin\DataGuru\StoreRequest;
+use App\Http\Requests\Admin\DataGuru\UpdateRequest;
 use App\Models\Guru;
 use App\Http\Controllers\Controller;
 use App\Models\GuruKelas;
@@ -11,7 +13,6 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use App\Models\User;
 use Spatie\Permission\Models\Role;
-use Spatie\Permission\Models\Permission;
 use App\Models\MataPelajaran;
 use App\Models\Kelas;
 
@@ -22,95 +23,72 @@ class DataGuruController extends Controller
         $roles = Role::all(); // Ambil semua role
         $opsiKelas = Kelas::all();
         $mataPelajarans = MataPelajaran::all(); // jangan dari Guru::with(), cukup ambil semua
-        $gurus = Guru::with(['user', 'kelas', 'mataPelajarans'])->get();
+        $gurus = Guru::with(['user.roles', 'kelas', 'mataPelajarans'])->get();
         return view('admin.dataguru.index', compact('gurus', 'roles', 'opsiKelas', 'mataPelajarans'));
     }
 
-    public function storeguru(Request $request)
+    public function storeguru(StoreRequest $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email|unique:users',
-            'password' => 'required|min:6',
-            'nip' => 'required|unique:gurus',
-            'jabatan' => 'required',
-            'mata_pelajaran' => 'required',
-            'jumlah_jam_mengajar' => 'required',
-            'jumlah_presensi' => 'required',
-            'role' => 'required',
-            'kelas' => 'required|array',
-            'kelas.*' => 'exists:kelas,id',
-            'mata_pelajaran' => 'required|array',
-            'mata_pelajaran.*' => 'exists:mata_pelajarans,id',
-        ]);
-
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => bcrypt($request->password),
-        ]);
-
-        $user->assignRole($request->role);
-
-        $guru = Guru::create([
-            'nip' => $request->nip,
-            'jabatan' => $request->jabatan,
-            'jumlah_jam_mengajar' => $request->jumlah_jam_mengajar,
-            'jumlah_presensi' => $request->jumlah_presensi,
-            'user_id' => $user->id, // hubungkan guru dengan user
-
-        ]);
-        // SIMPAN relasi ke tabel pivot guru_kelas
-        $guru->kelas()->sync($request->kelas);
-        $guru->mataPelajarans()->sync($request->mata_pelajaran);
+        DB::transaction(function() use($request){
+            $user = User::create([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+            ]);
+    
+            $user->assignRole($request->role);
+            $guru = Guru::create([
+                'nip' => $request->nip,
+                'jabatan' => $request->jabatan,
+                'jumlah_jam_mengajar' => $request->role === 'Guru' ? $request->jumlah_jam_mengajar : 0,
+                'jumlah_presensi' => $request->role === 'Guru' ? $request->jumlah_presensi : 0,
+                'user_id' => $user->id
+    
+            ]);
+    
+            if($request->role === "Guru"){
+                $guru->kelas()->sync($request->kelas);
+                $guru->mataPelajarans()->sync($request->mata_pelajaran);
+            }
+        });
 
         return redirect()->route('admin.dataguru.index')->with('success', 'Data guru berhasil ditambahkan.');
-
     }
 
-    public function update(Request $request, $id)
+    public function update(UpdateRequest $request, $id)
     {
         $guru = Guru::findOrFail($id);
         if (!$guru->user) {
             return redirect()->back()->with('error', 'User terkait guru ini tidak ditemukan.');
         }
+
         $user = $guru->user;
-
-        // $request->validate([
-        //     'nip' => 'required|unique:gurus,nip,' . $guru->id,
-        //     'nama' => 'required',
-        //     'jabatan' => 'required',
-        //     'mata_pelajaran' => 'required',
-        //     'jumlah_jam_mengajar' => 'required|numeric',
-        //     'jumlah_presensi' => 'required|numeric',
-        //     'email' => 'required|email|unique:users,email,' . $user->id,
-        //     'kelas_id' => 'required|exists:kelas,id',
-        //     'mata_pelajaran_id' => 'required|exists:mata_pelajarans,id',
-        // ]);
-
         DB::transaction(function () use ($request, $id, $user) {
+            $user->syncRoles($request->role);
             Guru::where('id', '=', $id)->update([
                 'nip' => $request->nip,
                 'jabatan' => $request->jabatan,
-                'jumlah_jam_mengajar' => $request->jumlah_jam_mengajar,
-                'jumlah_presensi' => $request->jumlah_presensi,
+                'jumlah_jam_mengajar' => $request->role === 'Guru' ? $request->jumlah_jam_mengajar : 0,
+                'jumlah_presensi' => $request->role === 'Guru' ? $request->jumlah_presensi : 0,
             ]);
 
             GuruKelas::where('guru_id', '=', $id)->delete();
             GuruMataPelajaran::where('guru_id', '=', $id)->delete();
-
-            foreach ($request->kelas as $row) {
-                GuruKelas::create([
-                    'guru_id' => $id,
-                    'kelas_id' => $row,
-                ]);
-            }
-
-            foreach ($request->mata_pelajaran as $row) {
-                GuruMataPelajaran::create([
-                    'guru_id' => $id,
-                    'mata_pelajaran_id' => $row,
-                ]);
+            if($request->role === "Guru"){
+    
+                foreach ($request->kelas as $row) {
+                    GuruKelas::create([
+                        'guru_id' => $id,
+                        'kelas_id' => $row,
+                    ]);
+                }
+    
+                foreach ($request->mata_pelajaran as $row) {
+                    GuruMataPelajaran::create([
+                        'guru_id' => $id,
+                        'mata_pelajaran_id' => $row,
+                    ]);
+                }
             }
 
             User::where('id', '=', $user->id)->update([
@@ -125,8 +103,11 @@ class DataGuruController extends Controller
 
     public function destroy($id)
     {
-        $guru = Guru::findOrFail($id);
-        $guru->delete();
+        DB::transaction(function() use($id){
+            $guru = Guru::findOrFail($id);
+            User::where('id', '=', $guru->user_id)->delete();
+            $guru->delete();
+        });
 
         return redirect()->route('admin.dataguru.index')->with('success', 'Data guru berhasil dihapus.');
     }
